+++ webview/src/types/index.ts
@@ -31,9 +31,15 @@ export interface IWorkbenchConfig {
     applyButtonText: string;
     backButtonText: string;
     collapseGroupAriaLabel: string;
+    connectToSourceText: string;
+    conditionalConnectToSourceText: string;
+    conditionalMenuAriaLabel: string;
+    conditionalRemoveConnectionText: string;
+    defaultHeaderText: string;
     expandGroupAriaLabel: string;
     nextButtonText: string;
     pageCountText: string;
+    unsupportedFieldTypeText: string;
     visibilityGroupName: string;
     visibilityToggleLabel: string;
     visibilityToggleOnText: string;